chore: split lexer struct into separate file
